pyproject.toml
[project]
name = "hr-intelligence"
version = "0.2.0"
description = "HR Intelligence Multi-Agent RAG"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb==1.5.2",
    "fastapi[standard]==0.115.12",
    "httpx==0.28.1",
    "langchain-core==1.2.16",
    "langchain-openai==1.1.10",
    "langgraph==1.0.7",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "pymupdf4llm==0.3.4",
    "python-docx==1.1.2",
    "python-dotenv==1.0.1",
    "rank-bm25==0.2.2",
    "sentence-transformers==3.4.1",
    "streamlit==1.43.0",
    "uvicorn[standard]==0.34.0",
]
